fix: ISEMPTY NPE in preview mode
Fixes: #2559
